+++ ui/src/configuration/components/LabelOverlayForm.tsx
@@ -3,31 +3,17 @@ import React, {PureComponent, ChangeEvent} from 'react'
 
 // Components
 import {
-  Stack,
   Button,
   Columns,
-  Alignment,
   ButtonType,
   ComponentSize,
   ComponentColor,
   ComponentStatus,
 } from '@influxdata/clockface'
-import {
-  Grid,
-  Form,
-  Input,
-  ComponentSpacer,
-  Label,
-  InputType,
-} from 'src/clockface'
-import LabelColorDropdown from 'src/configuration/components/LabelColorDropdown'
+import {Grid, Form, Input, Label, InputType, ColorPicker} from 'src/clockface'
 
 // Constants
-import {
-  CUSTOM_LABEL,
-  HEX_CODE_CHAR_LENGTH,
-  INPUT_ERROR_COLOR,
-} from 'src/configuration/constants/LabelColors'
+import {INPUT_ERROR_COLOR} from 'src/configuration/constants/LabelColors'
 const MAX_LABEL_CHARS = 50
 
 // Utils
@@ -43,14 +29,12 @@ interface Props {
   id: string
   name: string
   description: string
-  colorHex: string
+  color: string
+  onColorChange: (color: string, status?: ComponentStatus) => void
   onSubmit: () => void
   onCloseModal: () => void
   onInputChange: (e: ChangeEvent<HTMLInputElement>) => void
-  onColorHexChange: (colorHex: string) => void
-  onToggleCustomColorHex: (useCustomColorHex: boolean) => void
   onNameValidation: (name: string) => string | null
-  useCustomColorHex: boolean
   buttonText: string
   isFormValid: boolean
 }
@@ -61,14 +45,13 @@ export default class LabelOverlayForm extends PureComponent<Props> {
     const {
       id,
       name,
+      color,
       onSubmit,
       buttonText,
       description,
       onCloseModal,
       onInputChange,
-      onColorHexChange,
-      useCustomColorHex,
-      onToggleCustomColorHex,
+      onColorChange,
       isFormValid,
     } = this.props
 
@@ -83,13 +66,13 @@ export default class LabelOverlayForm extends PureComponent<Props> {
                     size={ComponentSize.Small}
                     name={this.placeholderLabelName}
                     description={description}
-                    colorHex={this.colorHexGuard}
+                    colorHex={this.colorGuard}
                     id={id}
                   />
                 </Form.Box>
               </Form.Element>
             </Grid.Column>
-            <Grid.Column widthSM={Columns.Seven}>
+            <Grid.Column widthSM={Columns.Twelve}>
               <Form.ValidationElement
                 label="Name"
                 value={name}
@@ -110,22 +93,6 @@ export default class LabelOverlayForm extends PureComponent<Props> {
                 )}
               </Form.ValidationElement>
             </Grid.Column>
-            <Grid.Column widthSM={Columns.Five}>
-              <Form.Element label="Color">
-                <ComponentSpacer
-                  align={Alignment.Left}
-                  stackChildren={Stack.Rows}
-                >
-                  <LabelColorDropdown
-                    colorHex={this.dropdownColorHex}
-                    onChange={onColorHexChange}
-                    useCustomColorHex={useCustomColorHex}
-                    onToggleCustomColorHex={onToggleCustomColorHex}
-                  />
-                  {this.customColorInput}
-                </ComponentSpacer>
-              </Form.Element>
-            </Grid.Column>
             <Grid.Column widthXS={Columns.Twelve}>
               <Form.Element label="Description">
                 <Input
@@ -135,6 +102,11 @@ export default class LabelOverlayForm extends PureComponent<Props> {
                   value={description}
                   onChange={onInputChange}
                 />
+              </Form.Element>
+            </Grid.Column>
+            <Grid.Column widthSM={Columns.Twelve}>
+              <Form.Element label="Color">
+                <ColorPicker color={color} onChange={onColorChange} />
               </Form.Element>
             </Grid.Column>
             <Grid.Column widthXS={Columns.Twelve}>
@@ -174,63 +146,17 @@ export default class LabelOverlayForm extends PureComponent<Props> {
     return name
   }
 
-  private get colorHexGuard(): string {
-    const {colorHex} = this.props
+  private get colorGuard(): string {
+    const {color} = this.props
 
-    if (validateHexCode(colorHex)) {
+    if (validateHexCode(color)) {
       return INPUT_ERROR_COLOR
     }
 
-    return colorHex
-  }
-
-  private get dropdownColorHex(): string {
-    const {colorHex, useCustomColorHex} = this.props
-
-    if (useCustomColorHex) {
-      return CUSTOM_LABEL.colorHex
-    }
-
-    return colorHex
+    return color
   }
 
   private handleNameValidation = (name: string): string | null => {
     return this.props.onNameValidation(name)
-  }
-
-  private get customColorInput(): JSX.Element {
-    const {colorHex, useCustomColorHex} = this.props
-
-    if (!useCustomColorHex) {
-      return null
-    }
-
-    return (
-      <Form.ValidationElement
-        label="Enter a Hexcode"
-        value={colorHex}
-        validationFunc={validateHexCode}
-      >
-        {status => (
-          <Input
-            type={InputType.Text}
-            value={colorHex}
-            placeholder="#000000"
-            onChange={this.handleCustomColorChange}
-            status={status}
-            autoFocus={true}
-            maxLength={HEX_CODE_CHAR_LENGTH}
-          />
-        )}
-      </Form.ValidationElement>
-    )
-  }
-
-  private handleCustomColorChange = (
-    e: ChangeEvent<HTMLInputElement>
-  ): void => {
-    const {onColorHexChange} = this.props
-
-    onColorHexChange(e.target.value)
   }
 }