+++ ui/src/configuration/components/UpdateLabelOverlay.tsx
@@ -4,9 +4,7 @@ import React, {Component, ChangeEvent} from 'react'
 // Components
 import LabelOverlayForm from 'src/configuration/components/LabelOverlayForm'
 import {OverlayContainer, OverlayBody, OverlayHeading} from 'src/clockface'
-
-// Utils
-import {validateHexCode} from 'src/configuration/utils/labels'
+import {ComponentStatus} from '@influxdata/clockface'
 
 // Types
 import {ILabel} from '@influxdata/influx'
@@ -23,7 +21,7 @@ interface Props {
 
 interface State {
   label: ILabel
-  useCustomColorHex: boolean
+  colorStatus: ComponentStatus
 }
 
 @ErrorHandling
@@ -33,26 +31,24 @@ class UpdateLabelOverlay extends Component<Props, State> {
 
     this.state = {
       label: props.label,
-      useCustomColorHex: false,
+      colorStatus: ComponentStatus.Default,
     }
   }
 
   public render() {
     const {onDismiss, onNameValidation} = this.props
-    const {label, useCustomColorHex} = this.state
+    const {label} = this.state
 
     return (
-      <OverlayContainer maxWidth={600}>
+      <OverlayContainer maxWidth={400}>
         <OverlayHeading title="Edit Label" onDismiss={onDismiss} />
         <OverlayBody>
           <LabelOverlayForm
             id={label.id}
             name={label.name}
             description={label.properties.description}
-            colorHex={label.properties.color}
-            onColorHexChange={this.handleColorHexChange}
-            onToggleCustomColorHex={this.handleToggleCustomColorHex}
-            useCustomColorHex={useCustomColorHex}
+            color={label.properties.color}
+            onColorChange={this.handleColorHexChange}
             onSubmit={this.handleSubmit}
             onCloseModal={onDismiss}
             onInputChange={this.handleInputChange}
@@ -66,10 +62,12 @@ class UpdateLabelOverlay extends Component<Props, State> {
   }
 
   private get isFormValid(): boolean {
-    const {label} = this.state
+    const {label, colorStatus} = this.state
 
     const nameIsValid = this.props.onNameValidation(label.name) === null
-    const colorIsValid = validateHexCode(label.properties.color) === null
+    const colorIsValid =
+      colorStatus === ComponentStatus.Default ||
+      colorStatus == ComponentStatus.Valid
 
     return nameIsValid && colorIsValid
   }
@@ -101,15 +99,14 @@ class UpdateLabelOverlay extends Component<Props, State> {
     }
   }
 
-  private handleColorHexChange = (color: string): void => {
+  private handleColorHexChange = (
+    color: string,
+    colorStatus: ComponentStatus
+  ): void => {
     const properties = {...this.state.label.properties, color}
     const label = {...this.state.label, properties}
 
-    this.setState({label})
-  }
-
-  private handleToggleCustomColorHex = (useCustomColorHex: boolean): void => {
-    this.setState({useCustomColorHex})
+    this.setState({label, colorStatus})
   }
 }